Split into Components
Split floor plan into further components
