+++ frontend/src/components/floorPlan/grid.js
@@ -1,0 +1,24 @@
+// 2D GRID FOR THE FLOOR PLAN
+
+import React from 'react';
+import { useHistory } from "react-router-dom";
+import { makeStyles } from "@material-ui/core/styles";
+import { Box, Button, Typography } from "@material-ui/core";
+import styled from "styled-components";
+
+const useStyles = makeStyles({
+    bold: {
+        fontWeight: 600
+      },
+})
+
+
+export default function Grid() {
+    const classes = useStyles();
+   
+    return (
+        <div>
+            {/* bryan */}
+        </div>
+   )
+}
